# Don't traverse directories that match file_ignore_regex and file_ignore_glob