@cachedout 2016-06-21
This looks good to me. I would also like @terminalmage to have a look.