@ctrlrsf 2016-07-25
Sorry for the delay, was traveling. Slice assignment was used to modify dirnames in place. From os.walk documentation:

https://docs.python.org/2/library/os.html#os.walk
> ... the caller can modify the dirnames list in-place (perhaps using del or slice assignment), and walk() will only recurse into the subdirectories whose names remain in dirnames; this can be used to prune the search

I thought using a slice assignment would be the most straight forward way to do so since its mentioned directly in the os.walk docs. If you prefer this be done another way, please let me know though.