@terminalmage 2016-07-26
@ctrlrsf Awesome, I had never come across this before. Thanks for clarifying, and helping me to learn something new today.